Refactor packages into internal/ except main and cmd, for clarity.
