+++ cmd/config.go
@@ -1,9 +1,8 @@
 package cmd
 
 import (
-	"github.com/davfive/gitspaces/v2/console"
-	"github.com/davfive/gitspaces/v2/gitspaces"
-
+	"github.com/davfive/gitspaces/v2/internal/config"
+	"github.com/davfive/gitspaces/v2/internal/console"
 	"github.com/spf13/cobra"
 )
 
@@ -13,7 +12,7 @@ var configCmd = &cobra.Command{
 	Short: "Open GitSpaces config file (yaml) in default editor",
 	Args:  cobra.NoArgs,
 	RunE: func(cmd *cobra.Command, args []string) error {
-		if err := gitspaces.User.OpenConfigFile(); err != nil {
+		if err := config.User.OpenConfigFile(); err != nil {
 			return console.Errorln("Failed to open GitSpaces config file:: %s", err)
 		}
 		return nil
