+++ internal/config/templates/gitspaces.cygwin.sh
