+++ internal/config/templates/gitspaces.ps1
