+++ internal/config/templates/gitspaces.scriptblock.ps1
