+++ internal/config/templates/gitspaces.sh
